Service log formatting quality of life improvements
Parameters which are interpolated into the formatted output are
no longer included in the trailing key/value parameter pairs, for
more concise output.

Parameters are automatically interpolated into the stackTrace
field, unsafe throwable messages are easier to follow in a
development environment.

Service log lines include the thread name on which events were
created, previously this data was not included in the formatted
output.

```diff
- ERROR [2019-05-09T15:32:37.692Z] ROOT: Hello, World! (0: World)
+ ERROR [2019-05-09T15:32:37.692Z] [main] ROOT: Hello, World!
```